Renamed package of resources in com.sun.xml.ws.config.management
- new package is same as the package of classes
