+++ wsit/metro-cm/metro-cm-api/src/main/java/module-info.java
@@ -1,4 +1,5 @@
 /*
+ * Copyright (c) 2026 Contributors to the Eclipse Foundation.
  * Copyright (c) 2022 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
@@ -12,6 +13,5 @@ module org.glassfish.metro.cm.api {
 
     requires transitive com.sun.xml.ws.rt;
 
-    exports com.sun.xml.ws.config.management;
     exports com.sun.xml.ws.metro.api.config.management;
 }
