+++ wsit/metro-cm/metro-cm-api/src/main/resources/com/sun/xml/ws/metro/api/config/management/Management_zh_CN.properties
